QUALITY: Modularise codebase
